New structure for 1D glaciers and climate (#26)
* Sleipnir root deprecated for Simulation working_dir

We have to stop using a Project-centered root_dir, since this fails when used in an installed packaged with the Pkg manager. Now we have a working_dir in Parameters.simulation.working_dir, which needs to be specified for each simulation

* Default OGGM ice thickness changed. == operator updated

* Minor fix on velocities flag

* Basic structure changed to accommodate 1D and 2D Glaciers and Climates

* A, C and n transfered from PhysicalParameters to Glacier

Allowing glacier-specific rheology tuning. Tests have also been updated with these changes