Adressed minor issues (#43)
@vivekag7 :

    Removed the printing of checking if glaciers are in GlaThiDa database.
    Improved name of helper function and made it useable for Python Objects and Julia Objects

Clarification on helper function:
Five tests in Huginn involve glaciers lacking a 'gdir'. The 'create_results()' function extracts longitude and latitude indirectly from the 'gdir'. This helper function safe_getproperty() makes sure that if the glacier has no gdir it still assigns a value (of zero) to the longitude and latitude. Acting as a workaround to accommodate glaciers without gdir.